+++ arg/option3.cc
@@ -1,12 +1,12 @@
 #include "arg.ih"
 
-unsigned Arg::option(unsigned idx, string *value, int optChar) const
+size_t Arg::option(size_t idx, string *value, int optChar) const
 {
     ISVMapIterator it = d_optv.find(optChar);
     if (it == d_optv.end())
         return 0;
 
-    unsigned ret = it->second.size();   // size of the value-vector
+    size_t ret = it->second.size();   // size of the value-vector
 
     if (idx < ret && value)             // if the idx is within range and
         *value = it->second[idx];       // value requested, return it.
