+++ configfile/removecomment.cc
@@ -2,7 +2,7 @@
 
 void ConfigFile::removeComment(string &line)
 {
-    unsigned pos;
+    size_t pos;
 
     if ((pos = line.find_first_of('#')) != string::npos)
         line.erase(pos);            // throw away beyond '#'
