+++ string/argv.cc
@@ -2,7 +2,7 @@
 
 char const **String::argv(std::vector<std::string> const &lines)
 {
-    unsigned idx = lines.size();
+    size_t idx = lines.size();
 
     char const **ret = new char const *[idx + 1];
     ret[idx] = 0;
